move store_linenoise_notch_CAR to pipelines
